requirements.txt
fastapi
uvicorn
python-dotenv
langchain-google-genai
langchain-community
langchain
langchain-huggingface
docx2txt
pydantic